Fix user 17 12 ---- Vers dev (#39)
* Update swagger

* Update form.json

* Update global.js

* Update User Routes & modules

* Update appart modules & UserService

+++ src/controller/UserController.js
@@ -1,14 +1,16 @@
-const UserService = require("../services/UserService");
-const baseUrl = require("../../config.json").baseUrl;
-const userService = new UserService();
+const UserService   = require("../services/UserService");
+const baseUrl       = require("../../config.json").baseUrl;
+const form          = require("../utils/form.json");
+const userService   = new UserService();
 
 class UserController {
+
     async authLogin(req, res) {
         try {
             const body = req.body;
             const resLogin = await userService.authLoginService(body.username, body.password);
             if (resLogin) res.status(200).json({message: "success", jwt: resLogin});
-            else res.status(403).json({message: "error"});
+            else res.status(403).json({error: "Impossible de se connecter"});
         } catch (error) {
             console.log(error);
             res.status(500).json({error: "Une erreur est survenue durant la connexion"});
@@ -32,7 +34,7 @@ class UserController {
             res.status(200).json({message: "success", users: usersData});
         } catch (error) {
             console.log(error);
-            res.status(500).json({error: "Une erreur est survenue durant la récupération de tous les users"});
+            res.status(500).json({error: "Une erreur est survenue durant la récupération des utilisateurs"});
         }
     }
 
@@ -40,31 +42,11 @@ class UserController {
         try {
             const userId = req.params.id;
             const userData = await userService.getUserService(userId);
-            res.status(200).json(userData);
+            if (!userData.error) res.status(200).json(userData);    // Si l'utilisateur existe
+            else res.status(404).json({error: userData.error});
         } catch (error) {
             console.log(error);
             res.status(500).json({error: "Une erreur est survenue durant la récupération du user"});
-        }
-    }
-
-    async getAllUsers(req, res) {
-        try {
-            const usersData = await userService.getAllUsersService();
-            res.status(200).json({message: "success", users: usersData});
-        } catch (error) {
-            console.log(error);
-            res.status(500).json({error: "Error during get all users"});
-        }
-    }
-
-    async getUserById(req, res) {
-        try {
-            const userId = req.params.id;
-            const userData = await userService.getUserService(userId);
-            res.status(200).json(userData);
-        } catch (error) {
-            console.log(error);
-            res.status(500).json({error: "Error during getting the user"});
         }
     }
 
@@ -73,7 +55,7 @@ class UserController {
             const registerData = req.body;
             const resRegister = await userService.registerService(registerData);
             if (resRegister) res.status(200).json({message: "success", jwt: resRegister});
-            else res.status(403).json({message: "error"});
+            else res.status(403).json({error: "Erreur durant l'inscription"});
         } catch (error) {
             console.log(error);
             res.status(500).json({error: "Une erreur est survenue durant l'inscription"});
@@ -85,10 +67,10 @@ class UserController {
             const userData = req.body;
             const resCreatedUser = await userService.createUser(userData);
             if (resCreatedUser) res.status(200).json(resCreatedUser);
-            else res.status(400).json({error: "Erreur durant la création"});
+            else res.status(400).json({error: form.errorCreateUser});
         } catch (error) {
             console.log(error);
-            res.status(500).json({error: "Erreur durant la création de l'utilisateur"});
+            res.status(500).json({error: form.errorCreateUser});
         }
     }
 
@@ -99,7 +81,7 @@ class UserController {
                 const resData = await userService.search(data);
                 res.status(200).json(resData);
             } else {
-                res.status(400).json({message: "Paramètres de recherche manquants", documentation: "https://doc"});
+                res.status(400).json({error: "Paramètres de recherche manquants"});
             }
         } catch (error) {
             console.log(error);
@@ -111,11 +93,11 @@ class UserController {
         try {
             const userIdDelete = req.params.id;
             const check = await userService.getUserService(userIdDelete);
-            if (check) {
+            if (check && check.id) {
                 userService.deleteUserById(userIdDelete);
                 res.status(200).json({message: "Suppression validée avec succès"});
             } else {
-                res.status(403).json({message: "error", cause: "Cet utilisateur n'existe pas"});
+                res.status(404).json({error: form.userNotFound});
             }
         } catch (error) {
             console.log(error);
@@ -131,33 +113,18 @@ class UserController {
                 isAdmin: req.user.isAdmin
             }
             const check = await userService.getUserService(data.userId);
-            if (check) {
+            if (check.id) { // Check si l'user existe
                 let result = await userService.patchUserById(data.userId, data);
                 if (result.link) res.status(200).json({message: "Modification réalisée avec succès !", redirect: result});
-                else res.status(403).json({message: "Erreur", cause: result});
+                else res.status(500).json({error: result});
             } else {
-                res.status(404).json({error: "L'utilisateur n'existe pas"});
+                res.status(404).json({error: form.userNotFound});
             }
         } catch (error) {
             console.log(error);
             res.status(500).json({error: "Une erreur est survenue durant la modification de l'utilisateur."});
         }
     }
-
-    /*async search(req, res) {
-        try {
-            const data = req.query;
-            if (Object.keys(data).length) {     // Renvoit un tableau des clés de l'objet, dont on récupère la taille
-                const resData = await userService.search(data);
-                res.status(200).json(resData);
-            } else {
-                res.status(400).json({message: "need a search query", documentation: "https://doc"});
-            }
-        } catch (error) {
-            console.log(error);
-            res.status(500).json({error: "Error during search"});
-        }
-    }*/
 }
 
 module.exports = UserController;
